+++ blueprints/codes/eurocode/en_1993_1_1_2022/chapter_8_ultimate_limit_state/formula_8_60.py
@@ -1,6 +1,6 @@
-"""Formula 8.60 from EN 1993-1-1:2025: Chapter 8 - Ultimate Limit State."""
+"""Formula 8.60 from EN 1993-1-1:2022: Chapter 8 - Ultimate Limit State."""
 
-from blueprints.codes.eurocode.en_1993_1_1_2025 import EN_1993_1_1_2025
+from blueprints.codes.eurocode.en_1993_1_1_2022 import EN_1993_1_1_2022
 from blueprints.codes.formula import Formula
 from blueprints.codes.latex_formula import LatexFormula, latex_replace_symbols
 from blueprints.type_alias import DIMENSIONLESS, MPA
@@ -11,7 +11,7 @@ class Form8Dot60ReducedYieldStrength(Formula):
     r"""Class representing formula 8.60 for the calculation of reduced yield strength for the shear area."""
 
     label = "8.60"
-    source_document = EN_1993_1_1_2025
+    source_document = EN_1993_1_1_2022
 
     def __init__(
         self,
@@ -20,7 +20,7 @@ class Form8Dot60ReducedYieldStrength(Formula):
     ) -> None:
         r"""Reduced yield strength calculation for the shear area when shear force exceeds 50% of plastic shear resistance [$MPa$].
 
-        EN 1993-1-1:2025 art.8.2.10(3) - Formula (8.60)
+        EN 1993-1-1:2022 art.8.2.10(3) - Formula (8.60)
 
         Parameters
         ----------
