--- blueprints/codes/eurocode/en_1993_1_1_2025/__init__.py
@@ -1,3 +1,0 @@
-"""EN 1993-1-1:2025."""
-
-EN_1993_1_1_2025 = "EN 1993-1-1:2025"
